+++ native/zohodesk_portal_core/lib/common/ZDPHomeEventListener.dart
@@ -1,0 +1,3 @@
+abstract class ZDPHomeEventListener {
+  void backButtonTapped();
+}